+++ tasks.py
@@ -27,7 +27,7 @@ def show_version(_):
 def lint(c, scope='all'):
     """Run PyLint."""
     if scope == 'all':
-        scope = 'c8y_api tests integration_tests samples'
+        scope = 'c8y_api c8y_tk tests integration_tests samples'
     c.run(f'pylint {scope}')